fix(engine): clear clippy warnings in TLS test helpers (green the engine gate)
The `cargo clippy --workspace --all-targets -- -D warnings` CI step failed on
three lints in `#[cfg(test)]` helpers added by the TLS-decrypt / HTTP-2 work.
They were clean under `--lib`, so they surfaced only under `--all-targets`:

- decrypt.rs: `repeat(x).take(n)` -> `repeat_n(x, n)` (manual_repeat_n)
- decrypt.rs: drop the unused `s_mac` binding — the server MAC key is
  intentionally unused (the test seals with the wrong MAC key)
- http2.rs: drop the redundant `sid as u32` cast (`sid` is already u32)

Fixed at the source (no #[allow]). `cargo fmt --all --check`, clippy with
`-D warnings`, and the ppcap-core TLS tests all pass locally.

+++ engine/crates/ppcap-core/src/tls/http2.rs
@@ -849,7 +849,7 @@ mod tests {
                     s.push(plen as u8);
                     s.push(ftype);
                     s.push(flags);
-                    s.extend_from_slice(&(sid as u32).to_be_bytes());
+                    s.extend_from_slice(&sid.to_be_bytes());
                     s.extend_from_slice(&payload);
                 }
                 // Random truncation to hit the mid-frame `break` paths.
